rename actions file to one more explicit
